renamed avocado to tang
