+++ demo/queries/user.js
@@ -1,4 +1,4 @@
-const builder = require('../../avocado/Builder')
+const builder = require('../../tang/Builder')
 const db = require('../connector')
 const users = require('../collections/users')
