+++ orango/ORM.js
@@ -1,8 +1,8 @@
-const Builder = require('../avocado/Builder')
+const Builder = require('../tang/Builder')
 const sortToAQL = require('./helpers/sortToAQL')
 const returnToAQL = require('./helpers/returnToAQL')
 const criteriaBuilder = require('./helpers/criteriaBuilder')
-const asyncForEach = require('../avocado/helpers/asyncForEach')
+const asyncForEach = require('../tang/helpers/asyncForEach')
 const setDefaultsToNull = require('./helpers/setDefaultsToNull')
 const DOC_VAR = 'doc'
 const EXPR = /"expr\([\s+]?([\w\s.+-]+)\)"/gi
