+++ orango/Orango.js
@@ -1,9 +1,9 @@
 const Connection = require('./Connection')
 const Schema = require('./Schema')
-const EventDispatcher = require('../avocado/EventDispatcher')
-const asyncForEach = require('../avocado/helpers/asyncForEach')
+const EventDispatcher = require('../tang/EventDispatcher')
+const asyncForEach = require('../tang/helpers/asyncForEach')
 const factory = require('./factory')
-const definePrivateProperty = require('../avocado/helpers/definePrivateProperty')
+const definePrivateProperty = require('../tang/helpers/definePrivateProperty')
 const CONSTS = require('./consts')
 
 class Orango {
