+++ tang/EventDispatcher.js
