+++ tang/factory.js
