+++ tang/helpers/clone.js
