+++ test/avocado.model.spec.js
@@ -1,5 +1,5 @@
 const expect = require('chai').expect
 
-describe('avocado model', () => {
+describe('tang model', () => {
 
 })
